feat(store): split blocking contract to separate classes (#68)
Before we release the first public version, I'd like to make room for a
non-blocking API in the future.

This is the first contract change to allow a suspending version of the
event store and its descending methods to be added in the future. I have
a local spike of a R2dbc PostgreSQL implementation, but since it's not
ready yet, I want to merge all the moves I had to do in order to have
both versions working in parallel. I'll also raise a draft of the
suspending version as soon as I have an end-to-end version working.